# Removing recursion in array.fs where it is may be worthwhile