I went through testing for opportunities to improve performance when recursive versions could be turned into explicit loops. For many cases it was a very small difference - 1% to 2%.  If that is deemed worthwhile I can roll those into this PR (would include functions like pick, and find, etc)

compareWith got a 20% performance improvement so seems like an clear win.  concat was a very small performance improvement, but I was also able to bring it down from 4 functions to just 2, so it is arguably simpler as well. I could pull this out though if the recursive implementation is preferred.

### compareWith

 Method |  Length |            Median |         StdDev | Scaled | Gen 0 | Gen 1 | Gen 2 | Bytes Allocated/Op |
------- |-------- |------------------ |--------------- |------- |------ |------ |------ |------------------- |
    **Old** |      **10** |        **23.7544 ns** |      **0.3000 ns** |   **1.00** |     **-** |     **-** |     **-** |               **0.00** |
    New |      10 |        20.6222 ns |      0.2443 ns |   0.87 |     - |     - |     - |               0.00 |
    **Old** |   **10000** |    **16,515.5425 ns** |    **137.0188 ns** |   **1.00** |     **-** |     **-** |     **-** |               **7.08** |
    New |   10000 |    13,171.2588 ns |    194.4002 ns |   0.80 |     - |     - |     - |               7.09 |
    **Old** | **1000000** | **1,741,685.5271 ns** | **46,344.7793 ns** |   **1.00** |     **-** |     **-** |  **2.00** |          **58,857.83** |
    New | 1000000 | 1,430,355.4282 ns | 20,364.5100 ns |   0.82 |     - |     - |  1.00 |          29,901.96 |


### concat (where length is how many arrays are being concatted, each array being 5 elements)

 Method | Length |             Median |          StdDev | Scaled | Gen 0 | Gen 1 | Gen 2 | Bytes Allocated/Op |
------- |------- |------------------- |---------------- |------- |------ |------ |------ |------------------- |
    **Old** |     **10** |        **313.5194 ns** |       **6.8561 ns** |   **1.00** |  **0.02** |     **-** |     **-** |              **96.07** |
    New |     10 |        303.6155 ns |       3.8994 ns |   0.97 |  0.02 |     - |     - |              92.10 |
    **Old** |   **5000** |    **157,129.6265 ns** |   **2,104.6524 ns** |   **1.00** |     **-** |     **-** |  **8.96** |          **41,074.92** |
    New |   5000 |    156,494.2465 ns |   4,158.5105 ns |   1.00 |     - |     - |  9.05 |          41,470.82 |
    **Old** | **500000** | **17,492,839.3981 ns** | **181,163.9916 ns** |   **1.00** | **75.00** | **75.00** | **96.00** |       **5,711,591.54** |
    New | 500000 | 17,198,734.2853 ns | 146,830.3554 ns |   0.98 | 81.67 | 77.59 | 65.34 |       4,449,222.19 |



### environment for tests

```ini

Host Process Environment Information:
BenchmarkDotNet=v0.9.8.0
OS=Microsoft Windows NT 6.2.9200.0
Processor=Intel(R) Core(TM) i7-4712HQ CPU 2.30GHz, ProcessorCount=8
Frequency=2240908 ticks, Resolution=446.2477 ns, Timer=TSC
CLR=MS.NET 4.0.30319.42000, Arch=64-bit RELEASE [RyuJIT]
GC=Concurrent Workstation
JitModules=clrjit-v4.6.1590.0

Type=SIMDBenchmark  Mode=Throughput  Platform=X64  
Jit=RyuJit  GarbageCollection=Concurrent Workstation  

```